book/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bookprintapi"
version = "0.1.0"
description = "BookPrintAPI Python SDK — 포토북 생성/주문 API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/sweet-book/bookprintapi-python-sdk"

[tool.setuptools.packages.find]
include = ["bookprintapi*"]
